% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{factorize}
\alias{factorize}
\title{Factorization method of given variables as string or list}
\usage{
factorize(x, dataset)
}
\arguments{
\item{x}{variable to be factorized as string}

\item{dataset}{dataset which you will be factoring from}
}
\value{
an updated dataset with the new factorized variables
}
\description{
Factor variables in the dataset by name
}
\details{
Factors given variable(s) of a dataset while providing an easy method for
assigning the the updated dataset to a dataset name
}
\examples{
# One covariate-------------------------------------------------------------
ghrelindat.new<-factorize("Hunger", dataset = ghrelindat)
# Note we set the output to a new variable
# We can also use integers to index columns
ghrelindat.new<-factorize(1, dataset = ghrelindat)

# More than one covariate---------------------------------------------------
ghrelindat.new <- factorize(c("Hunger","Eatingcond"),dataset = ghrelindat)

}
\author{
Kevin Duan
}
